+++ rust-src/concordium_base/src/common/cbor/primitives.rs
@@ -231,3 +231,145 @@ impl CborDeserialize for MapKey {
         }
     }
 }
+
+#[cfg(test)]
+mod test {
+    use super::*;
+    use crate::common::cbor::{cbor_decode, cbor_encode};
+
+    #[test]
+    fn test_u64() {
+        let value = 0u64;
+        let cbor = cbor_encode(&value).unwrap();
+        assert_eq!(hex::encode(&cbor), "00");
+        let value_decoded: u64 = cbor_decode(&cbor).unwrap();
+        assert_eq!(value_decoded, value);
+
+        let value = 1u64;
+        let cbor = cbor_encode(&value).unwrap();
+        assert_eq!(hex::encode(&cbor), "01");
+        let value_decoded: u64 = cbor_decode(&cbor).unwrap();
+        assert_eq!(value_decoded, value);
+
+        let value = 1230u64;
+        let cbor = cbor_encode(&value).unwrap();
+        assert_eq!(hex::encode(&cbor), "1904ce");
+        let value_decoded: u64 = cbor_decode(&cbor).unwrap();
+        assert_eq!(value_decoded, value);
+    }
+
+    #[test]
+    fn test_i64() {
+        let value = 0i64;
+        let cbor = cbor_encode(&value).unwrap();
+        assert_eq!(hex::encode(&cbor), "00");
+        let value_decoded: i64 = cbor_decode(&cbor).unwrap();
+        assert_eq!(value_decoded, value);
+
+        let value = 1i64;
+        let cbor = cbor_encode(&value).unwrap();
+        assert_eq!(hex::encode(&cbor), "01");
+        let value_decoded: i64 = cbor_decode(&cbor).unwrap();
+        assert_eq!(value_decoded, value);
+
+        let value = 2i64;
+        let cbor = cbor_encode(&value).unwrap();
+        assert_eq!(hex::encode(&cbor), "02");
+        let value_decoded: i64 = cbor_decode(&cbor).unwrap();
+        assert_eq!(value_decoded, value);
+
+        let value = -1i64;
+        let cbor = cbor_encode(&value).unwrap();
+        assert_eq!(hex::encode(&cbor), "20");
+        let value_decoded: i64 = cbor_decode(&cbor).unwrap();
+        assert_eq!(value_decoded, value);
+
+        let value = -2i64;
+        let cbor = cbor_encode(&value).unwrap();
+        assert_eq!(hex::encode(&cbor), "21");
+        let value_decoded: i64 = cbor_decode(&cbor).unwrap();
+        assert_eq!(value_decoded, value);
+
+        let value = 1230i64;
+        let cbor = cbor_encode(&value).unwrap();
+        assert_eq!(hex::encode(&cbor), "1904ce");
+        let value_decoded: i64 = cbor_decode(&cbor).unwrap();
+        assert_eq!(value_decoded, value);
+
+        let value = -1230i64;
+        let cbor = cbor_encode(&value).unwrap();
+        assert_eq!(hex::encode(&cbor), "3904cd");
+        let value_decoded: i64 = cbor_decode(&cbor).unwrap();
+        assert_eq!(value_decoded, value);
+    }
+
+    #[test]
+    fn test_bool() {
+        let value = false;
+        let cbor = cbor_encode(&value).unwrap();
+        assert_eq!(hex::encode(&cbor), "f4");
+        let value_decoded: bool = cbor_decode(&cbor).unwrap();
+        assert_eq!(value_decoded, value);
+
+        let value = true;
+        let cbor = cbor_encode(&value).unwrap();
+        assert_eq!(hex::encode(&cbor), "f5");
+        let value_decoded: bool = cbor_decode(&cbor).unwrap();
+        assert_eq!(value_decoded, value);
+    }
+
+    #[test]
+    fn test_bytes() {
+        let bytes = Bytes(vec![1, 2, 3, 4, 5]);
+
+        let cbor = cbor_encode(&bytes).unwrap();
+        assert_eq!(hex::encode(&cbor), "450102030405");
+        let bytes_decoded: Bytes = cbor_decode(&cbor).unwrap();
+        assert_eq!(bytes_decoded, bytes);
+    }
+
+    #[test]
+    fn test_bytes_exact_length() {
+        let bytes: [u8; 5] = [1, 2, 3, 4, 5];
+
+        let cbor = cbor_encode(&bytes).unwrap();
+        assert_eq!(hex::encode(&cbor), "450102030405");
+        let bytes_decoded: [u8; 5] = cbor_decode(&cbor).unwrap();
+        assert_eq!(bytes_decoded, bytes);
+
+        let err = cbor_decode::<[u8; 4]>(&cbor).unwrap_err().to_string();
+        assert!(err.contains("expected 4 bytes"), "err: {}", err);
+    }
+
+    /// Test where CBOR is not well-formed: Bytes length in header does not
+    /// match actual data. Test that we get an error and don't panic
+    #[test]
+    fn test_bytes_length_invalid() {
+        let cbor = hex::decode("58ff0102030405").unwrap();
+        cbor_decode::<[u8; 0xff]>(&cbor).expect_err("should give error");
+
+        let cbor = hex::decode("410102030405").unwrap();
+        cbor_decode::<[u8; 0x01]>(&cbor).expect_err("should give error");
+    }
+
+    #[test]
+    fn test_text() {
+        let text = "abcd";
+
+        let cbor = cbor_encode(&text).unwrap();
+        assert_eq!(hex::encode(&cbor), "6461626364");
+        let text_decoded: String = cbor_decode(&cbor).unwrap();
+        assert_eq!(text_decoded, text);
+    }
+
+    /// Test where CBOR is not well-formed: Text length in header does not match
+    /// actual data. Test that we get an error and don't panic
+    #[test]
+    fn test_text_length_invalid() {
+        let cbor = hex::decode("78ff61626364").unwrap();
+        cbor_decode::<String>(&cbor).expect_err("should give error");
+
+        let cbor = hex::decode("6161626364").unwrap();
+        cbor_decode::<String>(&cbor).expect_err("should give error");
+    }
+}
